backend/requirements/common.txt
aioredis==1.3.1
amqp==5.1.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
billiard==3.6.4.0
boto3==1.22.9
botocore==1.25.9
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.3.0
django-debug-toolbar==3.2.4
django-ninja==0.18.0
django-storages==1.12.3
django-timezone-field==4.2.3
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.5
httptools==0.3.0
httpx==0.21.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
kombu==5.2.4
msgpack==1.0.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
PyYAML==6.0
rfc3986==1.5.0
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.9
uvicorn==0.17.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
zope.interface==5.4.0
